rename: add faq component
